separated ui from replayer
