+++ packages/ui/src/main.js
@@ -1,4 +1,4 @@
-import App from './App.svelte';
+// import App from './App.svelte';
 import Embed from './Embed.svelte';
-
-export { App, Embed };
+import Replayer from "./Replayer.svelte"
+export {  Embed, Replayer };
